marker export fix
Allows markers to be exported properly even if there are multiple preceding the ones being exported. Basically ignores any markers outside of the selected export frames to accomplish this. 
Also includes a small fix for an issue I ran into where sometimes vertices would somehow be exported with NaN values, (possibly after being shrunk by bones). Not sure if this occurs in this version, but it can be ignored if not needed.